+++ src/packages/basemap-switcher/CHANGELOG.md
@@ -1,1 +1,14 @@
 # @open-pioneer/basemap-switcher
+
+## 0.1.0
+
+### Minor Changes
+
+-   1e7545c: Initial release.
+
+### Patch Changes
+
+-   Updated dependencies [bb2f27a]
+-   Updated dependencies [182da1c]
+    -   @open-pioneer/map@0.1.0
+    -   @open-pioneer/react-utils@0.1.0