+++ src/packages/map-navigation/CHANGELOG.md
@@ -1,1 +1,14 @@
 # @open-pioneer/map-navigation
+
+## 0.1.0
+
+### Minor Changes
+
+-   70b2381: Initial release
+
+### Patch Changes
+
+-   Updated dependencies [bb2f27a]
+-   Updated dependencies [182da1c]
+    -   @open-pioneer/map@0.1.0
+    -   @open-pioneer/react-utils@0.1.0